Move only execution integration test to integTest source set
